Update user global rating after a new rating was recorded

+++ src/main/java/fr/eql/teama/catalogue/entities/User.java
@@ -39,7 +39,7 @@ public class User {
 
     private String registrationNumber;
 
-    private Integer globalRating;
+    private Float globalRating;
 
     private String image;
